Problem: need way to provision a node
Apart from sharing posts over the protocol, we need a way to add
posts to a node. There are two main options I can see. One is to
monitor a directory (like FileMQ and Drops do). The other is to
send new posts via an API.

The advantage of sending via an API is that it lets us do things
like hook the Hydra API directly to a camera API, or other sources
of content.

Solution: add a provisioning API to hydra_server. This works with
actor messages:

    POST (subject, parent, string)
    POST FILE (subject, parent, MIME type, filename)
    POST DATA (subject, parent, MIME type, data, size)

Also, I've built a hydra_post class to abstract the storage format
for posts. This currently lets us write posts, though not yet read
them back.

In this commit, the hydra_ledger class is empty. It will server as
the main index for posts held by a node.