Problem: client destructor is not safe in all states
If client is expecting the reply from some other command, and then
the API calls the destructor, the state machine may end up reporting
a false internal error.

Solution: in the "expect goodbye ok" state, discard any other replies
from the server without complaining, or processing them. If the
client is receiving a post, that post will be abandoned.