Problem: server had half-finished credit mechanism
This was an abortive design to manage the flow of posts to specific
clients. However it was not solving a clear problem so I'd removed
it from the protocol and client. Dangling code in the server caused
an ACCESS REFUSED error after fetching 20 posts.

Solution: remove the dangling code from the server.

Fixes #66