Problem: hydra was not getting a valid hostname
It was using zsys_hostname() which is really not for networking use;
this method needs to be better documented. The correct way to get a
public facing hostname is to use zbeacon, so that we properly create
a socket on a valid interface, and then use that IP address.

Solution: get a hostname via zbeacon.